refactor: Move loadSeatCurrentlyInUse api from Store package to Utilization package
